backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.7.0
djangorestframework==3.17.1
sqlparse==0.5.5
tzdata==2026.1
